requirements.txt
bcrypt==4.1.3
black==24.4.2
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Bcrypt==1.0.1
greenlet==3.0.3
idna==3.7
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
